handling float data

+++ ravepay/utils.py
@@ -43,7 +43,7 @@ class RavepayAPI(object):
         result = response.json()
         if response.status_code == 200 and result["status"] == "success":
             data = result["data"]
-            if data["amount"] == int(kwargs.get("amount")):
+            if data["amount"] == float(kwargs.get("amount")):
                 return True, "Verification Successful"
             return False, data["amount"]
         if response.status_code >= 400:
